requirements.txt
fastapi
uvicorn
pillow
ultralytics
numpy
torch
torchvision
d